Use icons from icons folder for a few React components
Updates Checkbox, ModalStandard, and Select to use shared icons.

This is better because it means that icons won't get duplicated if two components are using the same icon. It also makes the icons a little easier to manage.

Fixes #39, fixes #41, and fixes #42.
