+++ packages/thumbprint-react/components/Select/index.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import * as tokens from '@thumbtack/thumbprint-tokens';
 import classNames from 'classnames';
 import styles from './index.module.scss';
+import { NavigationCaretDownSmall } from '../../icons/index.jsx';
 
 const iconColor = {
     disabled: tokens.tpColorGray,
@@ -65,21 +66,7 @@ const Select = ({
                 {children}
             </select>
 
-            {/**
-                Should be using Thumbprint Icons but blocked by:
-                https://github.com/thumbtack/thumbprint-archive/issues/687
-            */}
-            <svg
-                className={styles.caret}
-                fill={iconColor[uiState]}
-                xmlns="http://www.w3.org/2000/svg"
-                width="18"
-                height="18"
-                viewBox="0 0 18 18"
-                data-thumbprint-id="navigation-caret-down--small"
-            >
-                <path d="M14.646 6.764L9 13 3.311 6.725a1 1 0 0 1 1.342-1.482L9 10l4.285-4.699c.2-.187.435-.301.715-.301a1 1 0 0 1 1 1c0 .306-.151.537-.354.764z" />
-            </svg>
+            <NavigationCaretDownSmall className={styles.caret} fill={iconColor[uiState]} />
         </div>
     );
 };
