+++ packages/thumbprint-react/icons/index.jsx
@@ -1,23 +1,4 @@
-import React from 'react';
-
-const smallIconProps = {
-    xmlns: 'http://www.w3.org/2000/svg',
-    fill: 'currentColor',
-    width: '18',
-    height: '18',
-    viewBox: '0 0 18 18',
-};
-
-const NavigationCloseSmall = props => (
-    <svg {...props}>
-        <path
-            fillRule="evenodd"
-            d="M9.014 7.6L4.771 3.357a1 1 0 1 0-1.414 1.414L7.6 9.014l-4.243 4.243a1 1 0 1 0 1.414 1.414l4.243-4.243 4.243 4.243a1 1 0 0 0 1.414-1.414l-4.243-4.243 4.243-4.243a1 1 0 0 0-1.414-1.414L9.014 7.6z"
-        />
-    </svg>
-);
-
-NavigationCloseSmall.defaultProps = smallIconProps;
-
-// eslint-disable-next-line import/prefer-default-export
-export { NavigationCloseSmall };
+export { default as ContentActionsCheckSmall } from './content-actions-check-small.jsx';
+export { default as NavigationCaretDownSmall } from './navigation-caret-down-small.jsx';
+export { default as NavigationCloseSmall } from './navigation-close-small.jsx';
+export { default as NavigationCloseMedium } from './navigation-close-medium.jsx';
